// src/types/wajib-pajak-orang-pribadi-type.ts
export type TWajibPajakOrangPribadiFormSchema = {
  nama: string;
  email: string;
  status_pegawai: string;
  nip: string;
  kewarganegaraan: string;
  nama_negara: string;
  nik: string;
  nama_ktp: string;
  file_foto_ktp: string;
  nama_bank: string;
  no_rekening: string;
  nama_rekening: string;
  file_foto_bukti_rekening: string;
  ada_npwp: string;
  npwp: string;
  nama_npwp: string;
  provinsi_npwp: string;
  kabupaten_npwp: string;
  file_foto_npwp: string;
  no_passport: string;
  nama_passport: string;
  masa_berlaku_passport: Date;
  file_foto_passport: string;
  ptkp: string;
};

export type TWajibPajakOrangPribadiModel = {
  kode_wajib_pajak_orang_pribadi: string;
  nama: string;
  email: string;
  status_pegawai: string;
  nip?: string;
  kewarganegaraan: string;
  nama_negara: string;
  nik?: string;
  nama_ktp?: string;
  file_foto_ktp?: string;
  nama_bank: string;
  no_rekening: string;
  nama_rekening: string;
  file_foto_bukti_rekening?: string;
  ada_npwp: string;
  npwp?: string;
  nama_npwp?: string;
  provinsi_npwp?: string;
  kabupaten_npwp?: string;
  file_foto_npwp?: string;
  no_passport?: string;
  nama_passport?: string;
  masa_berlaku_passport?: string;
  file_foto_passport?: string;
  ptkp: string;
  disetujui: string;
  tanggal_input: Date;
};